Updated Application
The name of the tabs have been changed to (Basic Info, Skills & Stats, Inventory, Appearance, Backstory, Spell Sheet). Title on pivot has been changed to D&D Character Manager rather than D&D Util. This is due to it seeming a little unclear of what this application actually is. Added a logo for the taskbar.

I added logic to the all stat text boxes that allow for auto submit of data when losing focus. The values in these text boxes are submitted and put directly as the text block next to the text blocks.
Added logic behind the Modifier text blocks. The associated stat text block are parsed into an int subtracted by 10 and then divided by two. This allow for modifier to be set for each stat.

Inventory's weaponFinder Combo Box has content added with the name of each weapon from the Players handbook. There is no actual logic involved behind this. Only the names of each weapon.